--- components/MatchInfo.tsx
@@ -1,18 +1,0 @@
-import { useMatchIdList, useMatchList } from "@hooks/match";
-
-interface MatchInfoProps {
-  puuid: string;
-}
-
-const MatchInfo = ({puuid}: MatchInfoProps) => {
-  const { matchIdList, isMatchIdListLoading, isMatchIdListFetching } = useMatchIdList(puuid);
-  const { matchList, isMatchListLoading, isMatchListFetching } = useMatchList(matchIdList);
-
-  return (
-    <div>
-
-    </div>
-  );
-};
-
-export default MatchInfo;
